Refactor unittests and add sample data for a Vitodens222f
